+++ esp-idf-svc/CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.51.0] - 2025-01-15
+
 ### Breaking
 - New `embassy-time-driver` implementation compatible with `embassy-time-driver` V0.2 (#548)